Add support for isAuthorizedFor which accepts an object describing the full resource request
Updated the README.md to show the proper usage of the isAuthorizedFor(...) call
Ensured all tests pass. Fixed typo.

This commit enables the developer to provide their own context for the resource identifier instead of using the request path. It is the responsibility of the developer to ensure the request is built by the server without accepting the resource identifier from the client. To accept theresource identifier from the client would be insecure. The developer may use the request path if leveraging this library for middleware, or they may use context-sensitive identifiers and decouple themselves from the request path.

Updated & Rebased to:
Apply requested fixes & add unit testing